Fix mapping error

+++ aspnet-core/src/TACHYON.Application.Shared/DriverLocationLogs/dtos/DriverLocationLogDto.cs
@@ -7,7 +7,7 @@ namespace TACHYON.DriverLocationLogs.dtos
 {
     public class DriverLocationLogDto : EntityDto<long>
     {
-        public long? TripId { get; set; }
+        public int? TripId { get; set; }
         public double Longitude { get; set; }
         public double Latitude { get; set; }
     }
